requirements.txt
wtforms
werkzeug
flask
pymysql
jsonpath
poco
airtest
allure-pytest==2.8.6
allure-python-commons==2.8.6
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
comtypes==1.1.7
facebook-wda==0.2.1
hrpc==1.0.8
idna==2.8
importlib-metadata==0.23
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==7.2.0
mss==4.0.3
numpy==1.17.4
opencv-contrib-python==3.4.2.17
packaging==19.2
Pillow==6.2.1
pluggy==0.13.0
pocoui
py==1.8.0
pyparsing==2.4.5
pypiwin32==223
pytest==5.3.0
pytest-html==2.0.1
pytest-metadata==1.8.0
pywin32==227
pywinauto==0.6.3
requests==2.22.0
six==1.13.0
urllib3==1.25.7
wcwidth==0.1.7
websocket-client==0.48.0
XMind==1.2.0
zipp==0.6.0
pytest-rerunfailures